src/app/requirements.txt
streamlit
pydantic
python-dotenv
azure-ai-projects
azure-identity
uvicorn
gunicorn
pandas
numpy
pypdf
azure-core
azure-search-documents
openai
pyodbc
PyPDF2